requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
argh==0.26.2
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
awesome-mixins==0.15
Babel==2.3.4
billiard==3.3.0.23
blessings==1.6
bpython==0.16
celery==3.1.23
cffi==1.9.1
channels==1.1.6
-e git+https://github.com/DeveloperCielo/API-3.0-Python#egg=cieloApi3
colorama==0.3.7
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
coverage==4.1
cpf-cnpj==0.1.4
cryptography==1.7.1
curtsies==0.2.11
daphne==1.3.0
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-appconf==1.0.2
django-braces==1.9.0
django-celery==3.1.17
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.8
django-filter==1.0.4
django-filters==0.2.1
django-global-permissions==0.2.2
django-guardian==1.4.4
django-imagekit==3.3
django-rest-swagger==2.1.2
django-widget-tweaks==1.4.1
djangoajax==2.3.6
djangorestframework==3.6.3
docopt==0.6.2
drfdocs==0.0.10
easy-mask==1.3.4
easy-report==0.10
flower==0.9.1
future==0.16.0
gcloud==0.18.1
geopy==1.11.0
googleapis-common-protos==1.2.0
greenlet==0.4.12
gunicorn==19.7.1
httpie==0.9.3
httplib2==0.9.2
hyperlink==17.3.0
idna==2.2
incremental==17.5.0
invoke==0.19.0
isort==4.2.14
itypes==1.1.0
Jinja2==2.9.6
kombu==3.0.35
Markdown==2.6.6
MarkupSafe==1.0
model-mommy==1.2.6
msgpack-python==0.4.8
ndg-httpsclient==0.4.2
oauth2client==3.0.0
oauthlib==1.1.1
olefile==0.44
openapi-codec==1.3.2
packaging==16.8
pathtools==0.1.2
pilkit==1.1.13
Pillow==4.0.0
protobuf==3.0.0
psycopg2==2.6.2
pusher==1.6
py==1.4.34
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pydotplus==2.0.2
Pygments==2.1.3
pygraphviz==1.3.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
Pyrebase==3.0.9
pytest==3.1.2
pytest-django==3.1.2
pytest-watch==4.1.0
python-cards==1.3
python-decouple==3.0
-e git+https://github.com/juanjbrown/python-decouple-plus#egg=python_decouple_plus
python-resize-image==1.1.11
pytz==2016.4
PyYAML==3.12
redis==2.10.5
reportlab==3.3.0
requests==2.10.0
rsa==3.4.2
selenium==3.0.1
simplejson==3.11.1
six==1.10.0
splinter==0.7.4
sqlparse==0.2.3
sseclient==0.0.12
static3==0.7.0
tornado==4.2
Twisted==17.5.0
txaio==2.8.1
uritemplate==3.0.0
urllib3==1.17
watchdog==0.8.3
wcwidth==0.1.7
xmltodict==0.10.2
zope.interface==4.4.2
